% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics-package.R
\docType{package}
\name{acoustics-package}
\alias{acoustics}
\alias{acoustics-package}
\title{acoustics: Acoustics Analysis Package}
\description{
This package supports the visualization and analysis of acousitic data.
}
\keyword{internal}
