% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featExctract.R
\name{featExtract}
\alias{featExtract}
\title{An acoustics data analysis function}
\usage{
featExtract(wave)
}
\arguments{
\item{wave}{waveform}
}
\description{
This function allows you extract the entropy, fundamental frequency, and dominant frequency of acoustic waveforms.
}
\keyword{acoustics}
